added hidden to tabs
